+++ src/navigation/RouterConfig.js
@@ -4,12 +4,11 @@ import { LoginPage } from '../pages/auth/LoginPage';
 import { RegistrationPage } from '../pages/auth/RegistrationPage';
 import { RootPage } from '../pages/RootPage';
 import { NotFound } from '../pages/NotFound';
-// import { HomePage } from '../pages/todo-list/HomePage';
 import { CONFIRM, CONFIRM_BY_EMAIL, HOME_PAGE, LOGIN, REGISTRATION, ROOT_PAGE } from './paths.const';
 import PrivateRoute from './PrivateRoute';
 import { NotConfirmedPage } from '../pages/auth/Confirm/NotConfirmedPage';
 import { ConfirmedPage } from '../pages/auth/Confirm/ConfirmedPage';
-import { NewHomePage } from '../pages/todo-list/NewHomePage';
+import { HomePage } from '../pages/todo-list/HomePage';
 
 export const RouterConfig = () =>
   <Switch>
@@ -22,8 +21,7 @@ export const RouterConfig = () =>
 
     {/* private auth route to page with tasks */}
     <PrivateRoute path={ HOME_PAGE }>
-      {/* <Route component={ HomePage } /> */}
-      <Route component={ NewHomePage } />
+      <Route component={ HomePage } />
     </PrivateRoute>
 
     {/* for not found */}
